Improved code quality
