+++ src/pointers.cpp
@@ -441,7 +441,7 @@ namespace big
 		});
 
 		// Script VM
-		main_batch.add("VM", "E8 ? ? ? ? 48 85 FF 48 89 1D ? ? ? ?", [this](memory::handle ptr)
+		main_batch.add("VM", "E8 ? ? ? ? 48 85 FF 48 89 1D", [this](memory::handle ptr)
 		{
 			m_script_vm = ptr.add(1).rip().as<PVOID>();
 		});
